Cleaned up checking for mortality

+++ src/include/structures.h
@@ -700,8 +700,8 @@ typedef struct {
     double trans_deficit_leaf[2];
     double trans_deficit_canopy;
     int    not_dead;
-    int    death_year;
-    int    death_doy;
+    double death_year;
+    double death_doy;
 
 } canopy_wk;
